copy the data files to the containers instead of using volume binds
