+++ ingestion/tests/integration/sql_server/conftest.py
@@ -19,6 +19,8 @@ from metadata.generated.schema.entity.services.databaseService import (
     DatabaseServiceType,
 )
 
+from _openmetadata_testutils.helpers.docker import copy_dir_to_container
+
 
 @pytest.fixture(scope="module")
 def mssql_container(tmp_path_factory):
@@ -42,9 +44,9 @@ GO
         """
         )
 
-    container.volumes = {str(data_dir): {"bind": "/data"}}
     with try_bind(container, 1433, 1433) as container:
         docker_container = container.get_wrapped_container()
+        copy_dir_to_container(str(data_dir), docker_container, "/data")
         res = docker_container.exec_run(
             [
                 "/opt/mssql-tools/bin/sqlcmd",
